refactor: 도메인 용어 재정립에 따라 코드에 반영
