--- src/main/java/HorizonLine.java
@@ -1,3 +1,0 @@
-public enum HorizonLine {
-    EXIST, EMPTY
-}
