+++ src/main/java/Stick.java
@@ -1,0 +1,3 @@
+public enum Stick {
+    FILLED, EMPTY
+}
